hwui: skip draw empty frame
On some hardare platform, when the buffer has been abandoned, the
query width and height of surface is random, the frame is empty, so
there is not need to draw the empty frame. We should skip it, if not
the hwui will abort for some glerror. This may be a hardware platform
issue, but improve the hwui stability is also needed. This issue can
be found on ARM/IMG GPU. For more crash information, we can check
the bug: https://code.google.com/p/android/issues/detail?id=224335

Bug: 224335
Test: manual

Change-Id: I6aac506c52aa88642541803e77a76e7bdd6aa3eb

Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>